requirements.txt
pillow==11.3.0
pandas==2.3.2
pyarrow==21.0.0
numpy==2.0.2
torch==2.9.1
torchvision==0.24.1
transformers==4.57.6
qwen-vl-utils==0.0.14
vllm==0.16.0
tqdm==4.67.1
accelerate==1.12.0
tokenizers==0.22.2
safetensors==0.7.0
openpyxl==3.1.5
huggingface_hub==0.36.2
datasets==4.1.1
openai==2.24.0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.7.2
